refactor(test): split pdf overlay tests by concern
