+++ src/app/overlays/pdf/tests/mod.rs
@@ -1,0 +1,6 @@
+mod geometry;
+mod helpers;
+mod pipeline;
+mod protocols;
+mod session;
+mod static_images;
